Adjust the migration categorization algorithm (#94)
To simplify the output and mental model of the categorization needed for safemigrate, eliminate the protected category by combining it with the delayed category, and expand the blocked category to include migrations that are only blocked because they are behind a blocked migration.

This allows for a clearer conceptual linearization of safemigrate's categorizations of migrations, so that they now flow from ready to delayed to blocked through the dependency graph.